// src/mitre-attack-coverage/model/coverage_overview/rule.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

import type { RuleObjectId } from '../../shims/api_types';

export interface CoverageOverviewRule {
  id: RuleObjectId; // rule SO's ids (not ruleId)
  name: string;
}